Fix: public responseParser (#1154)
* fix: enable instance parser change

* Update __tests__/rpcProvider.test.ts


* fix: export ResponseParser and RPCResponseParser

* chore: noob

---------
